+++ gdsfactory/port.py
@@ -108,12 +108,13 @@ class Port(PortPhidl):
 
     @classmethod
     def validate(cls, v):
-        """For pydantic assumes Port is valid if:
+        """For pydantic assumes Port is valid if has a name and a valid type
 
         - has a name
         """
         assert v.name, f"Port has no name, got `{v.name}`"
         # assert v.assert_on_grid(), f"port.midpoint = {v.midpoint} has off-grid points"
+        assert isinstance(v, Port), f"TypeError, Got {type(v)}, expecting Port"
         return v
 
     @property